Сделал тест компонента Main. Пришлось переделать работу с контекстом
